Fix conditional attribute rendering for Ruby expressions
When a hash attribute value is a Ruby expression (wrapped in <%= %>),
it should be rendered conditionally rather than always present.

HAML convention: if an attribute value is nil/false, the attribute
should not be rendered at all.

Changes:
- Detect when attribute value starts with <%= and ends with %>
- Convert to conditional ERB: <%= ' attr="attr"' if (expr) %>
- This properly omits the attribute when expr is nil/false

Example:
  HAML: %option{ selected: (time == value) ? "selected" : nil }
  Before: <option selected="<%= (time == value) ? "selected" : nil %>">
  After:  <option <%= ' selected="selected"' if ((time == value) ? "selected" : nil) %>>

This fixes the bug where boolean attributes would always be present
even when the expression evaluated to nil.